move to package org.apache.spark.sql.dialect
